Refactor and add support for custom formatters
